requirements.txt
# Video Processing
yt-dlp==2023.12.30
opencv-python==4.9.0.80
moviepy==1.0.3
pillow==10.2.0

# AI/ML Libraries
google-generativeai==0.3.2
openai-whisper==20231117
torch==2.1.2
transformers==4.37.2
sentence-transformers==2.2.2

# Web Framework
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.30.0

# Database & Storage
chromadb==0.4.22

# Utilities
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pytesseract==0.3.10
numpy==1.26.3
pandas==2.2.0
matplotlib==3.8.2
seaborn==0.13.1

# Document Generation
reportlab==4.0.9
markdown==3.5.2
markdown2==2.4.12

# Additional utilities
tqdm==4.66.1
typing-extensions==4.9.0
pydantic==2.5.3
